fix: free `to` Tensor in Matmul files. (#210)
* feat: change Tenosr::count to uint64_t.
* fix: free `to` Tensor in `Matmul` files.

+++ include/Types.hpp
@@ -330,7 +330,7 @@ static string DataTypeName(DataType dataType) {
         return "Unknown";
     }
 }
-static size_t DataTypeSize(DataType dtype, int count = 1) {
+static size_t DataTypeSize(DataType dtype, uint64_t count = 1) {
     switch (dtype) {
     case MLLM_TYPE_F32:
         return sizeof(float) * count;
